requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.6
ConfigArgParse==1.7
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.1
django-mysql==4.9.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
emoji==2.5.1
exceptiongroup==1.1.1
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.7.0
geventhttpclient==2.0.9
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.3.1
locust==2.15.1
MarkupSafe==2.1.3
msgpack==1.0.5
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
openapi-codec==1.3.2
outcome==1.2.0
packaging==23.1
pandas==2.0.2
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
pyproj==3.5.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.30.0
roundrobin==0.0.4
scikit-learn==1.3.0
selenium==4.10.0
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
webdriver-manager==3.8.6
Werkzeug==2.3.6
wsproto==1.2.0
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
